requirement.txt
tqdm
psycopg2-binary
numpy
scipy
pandas
django
requests
nltk
gensim
stop-words